% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colDef_agg.R
\name{colDef_agg}
\alias{colDef_agg}
\title{Generate column definitions for aggregating specified columns by a specified metric in a Reactable table}
\usage{
colDef_agg(cols, rm_footer = T, agg_fun = "sum")
}
\arguments{
\item{cols}{A character vector specifying the column names to be aggregated.}

\item{rm_footer}{A logical value indicating whether to remove the footer from the column definitions.
Default is \code{TRUE}.}

\item{agg_fun}{A character string specifying the aggregation function to be applied.
Default is \code{"sum"}. Other possible values include functions available
in the \code{dplyr} package such as \code{"mean"}, \code{"median"}, etc.}
}
\value{
A list of column definitions suitable for use in a Reactable table with specified
        columns aggregated by the specified metric.
}
\description{
The \code{colDef_agg} function generates column definitions for aggregating specified columns
by a specified metric in a Reactable table. It creates column definitions based on provided column
names, a metric aggregation function, and whether to remove the footer from the column definitions.
}
\examples{

library(dplyr)
library(tibble)
library(magrittr)
library(reactable)

mtcars \%>\%
  rownames_to_column("cars") \%>\%
  mutate(cars = gsub(" .*", "\\\\1", cars)) \%>\%
  reactable(
    defaultColDef = colDef(footerStyle = list(fontWeight = "bold")),
    groupBy = "cars",
    columns = combined_named_lists(
      colDef_agg(cols = c("am", "gear", "carb"), agg_fun = "sum")
    )
  )
}
