% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colDef_colWidth_robust.R
\name{colDef_colWidth_robust}
\alias{colDef_colWidth_robust}
\title{Define column widths for attributes in a reactable table}
\usage{
colDef_colWidth_robust(
  cols,
  minWidth = NA,
  maxWidth = NA,
  width = NA,
  wrap = F
)
}
\arguments{
\item{cols}{A character vector specifying the attributes for which column widths are defined.}

\item{minWidth}{Minimum width for the columns. Default is NA.}

\item{maxWidth}{Maximum width for the columns. Default is NA.}

\item{width}{Width for the columns. Default is NA.}

\item{wrap}{Logical indicating whether to wrap text in the columns. Default is FALSE.}
}
\value{
A list of column definitions for the specified attributes, with width settings applied.
}
\description{
This function allows you to define column widths for different attributes in a reactable table.
}
\examples{

rownames_to_column("cars") \%>\%
  reactable(
    defaultColDef = colDef(footerStyle = list(fontWeight = "bold")),
    groupBy = "cars",
    columns = combined_named_lists(
      colDef_colWidth_robust(cols = c("cars", "mpg"), minWidth = 400)
      # ,colDef_hide(cols = c("vs", "am", "gear", "carb", "wt"))
    )
  )

}
