% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colDef_filter_select.R
\name{colDef_filter_select}
\alias{colDef_filter_select}
\title{Generate column definitions with filter dropdown menus for specified columns in a Reactable table}
\usage{
colDef_filter_select(cols, id)
}
\arguments{
\item{cols}{A character vector specifying the column names for which the filter dropdown menus
column definitions are to be generated.}

\item{id}{A character string specifying the unique ID for the Reactable table. This ID should be
specified both in the function call and in the Reactable table definition using the
`elementId` parameter and setting `filterable = TRUE`.}
}
\value{
A list of column definitions suitable for use in a Reactable table with filter dropdown menus.
}
\description{
The \code{colDef_filter_select} function generates column definitions with filter dropdown menus
for specified columns in a Reactable table. It creates column definitions based on provided column
names and a unique ID for the Reactable table.
}
\examples{

# Set the unique ID for the Reactable table
id <- "temp_id"

mtcars \%>\%
  rownames_to_column("cars") \%>\%
  reactable(
    defaultColDef = colDef(footerStyle = list(fontWeight = "bold")),
    columns = combined_named_lists(
      colDef_filter_select(cols = c("cars"), id = id)
    ),
    elementId = id, filterable = TRUE
  )
}
