% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colDef_filter_sort_labels.R
\name{colDef_filter_sort_labels}
\alias{colDef_filter_sort_labels}
\title{Reorder Vector to Move Null Values to End}
\usage{
colDef_filter_sort_labels(vec)
}
\arguments{
\item{vec}{A character vector containing elements to be reordered.}
}
\value{
A reordered character vector with 'NULL', 'Null', and 'null' values at the end.
}
\description{
This function reorders a vector such that elements which are 'NULL', 'Null',
or 'null' appear at the end of the vector. It is used inside another
`colDef` function to help sort filter labels so that iterations of null
appear last in the filter drop-down.
}
\examples{
vec <- c("A", "null", "B", "D")
new_vec <- colDef_filter_sort_labels(vec)
print(new_vec)

}
