% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colDef_hide.R
\name{colDef_hide}
\alias{colDef_hide}
\title{Generate column definitions to hide specified columns in a Reactable table}
\usage{
colDef_hide(cols)
}
\arguments{
\item{cols}{A character vector specifying the column names to be hidden.}
}
\value{
A list of column definitions suitable for use in a Reactable table.
}
\description{
The \code{colDef_hide} function generates column definitions to hide specified columns
in a Reactable table without removing them from the data frame object. It creates column
definitions based on provided column names.
}
\examples{

mtcars \%>\%
  rownames_to_column("cars") \%>\%
  reactable(
    defaultColDef = colDef(footerStyle = list(fontWeight = "bold")),
    columns = combined_named_lists(
      colDef_hide(cols = c("vs", "am", "gear", "carb", "wt"))
    )
  )
}
