% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colDef_minwidth.R
\name{colDef_minwidth}
\alias{colDef_minwidth}
\title{Generate column definitions with minimum width for specified columns in a Reactable table}
\usage{
colDef_minwidth(cols, width)
}
\arguments{
\item{cols}{A character vector specifying the column names for which the minimum width
column definitions are to be generated.}

\item{width}{An integer specifying the minimum width for the specified columns.}
}
\value{
A list of column definitions suitable for use in a Reactable table.
}
\description{
The \code{colDef_minwidth} function generates column definitions with a minimum width
for specified columns in a Reactable table. It creates column definitions based on
provided column names and the desired minimum width.
}
\examples{
mtcars \%>\%
  rownames_to_column("cars") \%>\%
  reactable(
    defaultColDef = colDef(footerStyle = list(fontWeight = "bold")),
    columns = combined_named_lists(
      colDef_minwidth(cols = c("mpg", "cyl"), width = 300)
    )
  )
}
