% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colDef_sticky.R
\name{colDef_sticky}
\alias{colDef_sticky}
\title{Generate sticky column definitions for a Reactable table}
\usage{
colDef_sticky(cols, side = "left")
}
\arguments{
\item{cols}{A character vector specifying the column names for which the sticky column definitions are to be generated.}

\item{side}{A character string specifying the side to which the columns should stick. Default is "left".}
}
\value{
A list of column definitions suitable for use in a Reactable table.
}
\description{
The \code{colDef_sticky} function generates column definitions with sticky behavior
for a Reactable table. It creates column definitions based on provided column names
and the side to which the columns should stick.
}
\examples{
library(reactable)
library(dplyr)

mtcars \%>\%
  rownames_to_column("cars") \%>\%
  reactable(
    defaultColDef = colDef(footerStyle = list(fontWeight = "bold")),
    columns = combined_named_lists(
      colDef_sticky(cols = c("cars", "mpg"))
    )
  )

# This example creates a Reactable table with sticky columns
# using the \code{colDef_sticky} function.
}
