% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colDef_urlLink.R
\name{colDef_urlLink}
\alias{colDef_urlLink}
\title{Generate column definitions for hyperlinked URLs}
\usage{
colDef_urlLink(cols, link_text)
}
\arguments{
\item{cols}{A character vector specifying the column names for which the hyperlink column definitions are to be generated.}

\item{link_text}{A character string specifying the text to be displayed as the link.}
}
\value{
A list of column definitions suitable for use in a Reactable table.
}
\description{
The \code{colDef_urlLink} function generates column definitions for hyperlinked URLs
in a Reactable table. It creates column definitions based on provided column names and
a text to display as the link.
}
\examples{

mtcars \%>\%
  mutate(url_raw = "www.google.com",
         url_names = rownames(.)) \%>\%
  remove_rownames() \%>\%
  reactable(
    defaultColDef = colDef(footerStyle = list(fontWeight = "bold")),
    columns = combined_named_lists(
      colDef_urlLink(cols = "url_raw", link_text = "Example text to link")
    )
  )

}
