% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_named_lists.R
\name{combined_named_lists}
\alias{combined_named_lists}
\title{Combine multiple column definitions into a single named list for use in a Reactable table}
\usage{
combined_named_lists(...)
}
\arguments{
\item{...}{Multiple sets of column definitions generated by functions like \code{colDef_sticky},
\code{colDef_hide}, \code{colDef_minwidth}, etc.}
}
\value{
A named list of combined column definitions suitable for use in a Reactable table.
}
\description{
The \code{combined_named_lists} function combines multiple column definitions generated by
various functions into a single named list suitable for use in a Reactable table. It takes
multiple sets of column definitions as input and combines them into a single named list.
}
\examples{

mtcars \%>\%
  rownames_to_column("cars") \%>\%
  reactable(
    defaultColDef = colDef(footerStyle = list(fontWeight = "bold")),
    groupBy = "cars",
    columns = combined_named_lists(
      colDef_sticky(cols = c("cars", "mpg"))
      ,colDef_hide(cols = c("vs", "am", "gear", "carb", "wt"))
    )
  )
}
