% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtrn_cols.R
\name{rtrn_cols}
\alias{rtrn_cols}
\title{Specify column definitions for instances of many columns at once without specifying them individually}
\usage{
rtrn_cols(data, words, pretty = FALSE, exclude = FALSE, sort = TRUE)
}
\arguments{
\item{data}{The data frame from which column names are extracted.}

\item{words}{A singular string input containing column names, where "|" signifies "or".}

\item{pretty}{A logical value indicating whether to return the column names in a pretty format. Default is \code{FALSE}.}

\item{exclude}{A logical value indicating whether to exclude the specified columns. Default is \code{FALSE}.}

\item{sort}{A logical value indicating whether to sort the column names alphabetically. Default is \code{TRUE}.}
}
\value{
A vector of column names specified by the input string.
}
\description{
The \code{rtrn_cols} function simplifies the process of specifying column definitions for instances
of many columns at once without specifying them individually. It takes a singular string input as the
source of column names, where the "|" operator signifies "or". If \code{exclude = TRUE}, it returns
all columns that are not specified in the input string.
}
\examples{

# Create example data
temp_data <- mtcars \%>\%
  rownames_to_column("cars") \%>\%
  mutate(cars = gsub(" .*", "\\\\1", cars))

print(rtrn_cols(data = temp_data, words = "cars", exclude = T))
print(rtrn_cols(data = temp_data, words = "cars", exclude = F))

# Create a Reactable table with specified column definitions
temp_data \%>\%
  reactable(
    defaultColDef = colDef(footerStyle = list(fontWeight = "bold")),
    groupBy = "cars",
    columns = combined_named_lists(
      colDef_agg(cols = rtrn_cols(data = temp_data, words = "cars", exclude = TRUE)),
      colDef_hide(cols = rtrn_cols(data = temp_data, words = "gear|am", exclude = FALSE))
    )
  )
}
